Reimplement the old JSON endpoints as API calls
Whyyy does this have to be so goddamn convoluted 😩 Just give me
direct access to the result array instead of forcing me to do this
addValue() garbage. Then your code would *maybe* be fast enough to
deliver more *actual content* to the caller, with *no batching*!
Imagine that! Wouldn't that be a wonderful world? 🤔